+++ src/demo/views/Components.vue
@@ -13,6 +13,7 @@
     <Tag class="mt-10" />
     <TagGroup class="mt-10" />
     <Avatar class="mt-10" />
+    <Tooltip class="mt-10" />
     <Selects class="mt-10" />
     <TextAreas class="mt-10" />
   </div>
@@ -32,6 +33,7 @@ import TagGroup from '../sections/TagGroup.vue'
 import Tag from '../sections/Tag.vue'
 import Popups from '../sections/Popups.vue'
 import Sliders from '../sections/Sliders.vue'
+import Tooltip from '../sections/Tooltip.vue'
 import Selects from '../sections/Selects.vue'
 import TextAreas from '../sections/TextAreas.vue'
 
@@ -50,7 +52,8 @@ import TextAreas from '../sections/TextAreas.vue'
     TabSection,
     Tag,
     TagGroup,
-    TextAreas
+    TextAreas,
+    Tooltip
   }
 })
 export default class App extends Vue {}